Refactor: server/api/service/apidoc -> server/api/service/openapi
